chore: remove identify function from lib
